backend/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2==2.9.9
pandas==2.2.0
scikit-learn==1.2.2
tensorflow-cpu==2.15.0